+++ src/raytracer/compositor/surfaceiterator.rs
@@ -2,12 +2,12 @@ use raytracer::compositor::{ColorRGBA, SurfaceFactory};
 
 
 pub struct SurfaceIterator {
-    x_delta: uint,
-    x_off: uint,
-    y_delta: uint,
-    y_off: uint,
-    parent_width: uint,
-    parent_height: uint,
+    x_delta: usize,
+    x_off: usize,
+    y_delta: usize,
+    y_off: usize,
+    parent_width: usize,
+    parent_height: usize,
     background: ColorRGBA<u8>,
 }
