+++ src/util/mod.rs
@@ -4,7 +4,7 @@ use std::num::Float;
 pub mod export;
 pub mod import;
 
-pub fn print_progress(noun: &str, start_time: ::time::Timespec, done: uint, total: uint) {
+pub fn print_progress(noun: &str, start_time: ::time::Timespec, done: usize, total: usize) {
     let remaining_jobs = total - done;
     let progress: f64 = 100f64 * done as f64 / total as f64;
     let current_time = ::time::get_time().sec;
@@ -25,8 +25,8 @@ pub fn print_progress(noun: &str, start_time: ::time::Timespec, done: uint, tota
     }
 }
 
-fn make_progress_bar(ratio: f64, length: uint) -> String {
-    let filled = (ratio * length as f64).round() as uint;
+fn make_progress_bar(ratio: f64, length: usize) -> String {
+    let filled = (ratio * length as f64).round() as usize;
     let mut bar: String = repeat('|').take(filled).collect();
 
     for _ in range(0, length - filled) {
